Add support for My Tag auto playlist
* Remove GENRE_TAG_DELIMITER option
* Rename GENERATE_GENRE_PLAYLISTS to GENERATE_TAGS_PLAYLISTS
* Rename GENERATE_GENRE_PLAYLISTS_REMAINDER to
GENERATE_TAGS_PLAYLISTS_REMAINDER
* Rename generate_genre_playlists.json to generate_tags_playlists.json
* Removed get_genres module and related options
* generate_tags_playlists module now create two auto-playlists:
  - My Tags
  - Genres
