+++ src/djtools/configs/config.json
@@ -18,11 +18,8 @@
     "DOWNLOAD_XML": false, 
     "UPLOAD_MUSIC": false, 
     "UPLOAD_XML": false, 
-    "GET_GENRES": false,
-    "GENRE_EXCLUDE_DIRS": [],
-    "GENRE_TAG_DELIMITER": "/",
-    "GENERATE_GENRE_PLAYLISTS": false,
-    "GENERATE_GENRE_PLAYLISTS_REMAINDER": "folder",
+    "GENERATE_TAGS_PLAYLISTS": false,
+    "GENERATE_TAGS_PLAYLISTS_REMAINDER": "folder",
     "GENERATE_GENRE_PLAYLISTS_PURE": [],
     "CHECK_TRACK_OVERLAP": false,
     "CHECK_TRACK_OVERLAP_FUZZ_RATIO": 80,
